Simplify cmd package modules
