+++ .vscode/launch.json
@@ -22,8 +22,7 @@
       "args": ["daemon", "start"],
       "env": {
         "PATH": "/usr/local/go/bin:/usr/bin:/usr/sbin",
-        "CHECKPOINT_SERVICE_URL": "http://localhost:1324",
-        "CEDANA_GPU_ENABLED": "true"
+        "CHECKPOINT_SERVICE_URL": "http://localhost:1324"
       }
     },
     {
