+++ build-start-daemon.sh
@@ -7,7 +7,7 @@ SERVICE_FILE="/etc/systemd/system/$APP_NAME.service"
 USER=$(whoami)
 CEDANA_GPU_ENABLED=${CEDANA_GPU_ENABLED:-0}
 CEDANA_OTEL_ENABLED=${CEDANA_OTEL_ENABLED:-0}
-GPU_CONTROLLER_PATH="/usr/local/bin/gpu-controller"
+CEDANA_GPU_CONTROLLER_PATH="/usr/local/bin/gpu-controller"
 CEDANA_PROFILING_ENABLED=${CEDANA_PROFILING_ENABLED:-0}
 CEDANA_IS_K8S=${CEDANA_IS_K8S:-0}
 CEDANA_GPU_DEBUGGING_ENABLED=${CEDANA_GPU_DEBUGGING_ENABLED:-0}
@@ -43,7 +43,7 @@ Description=Cedana Checkpointing Daemon
 [Service]
 Environment=USER=$USER
 Environment=CEDANA_GPU_ENABLED=$CEDANA_GPU_ENABLED
-Environment=GPU_CONTROLLER_PATH=$GPU_CONTROLLER_PATH
+Environment=CEDANA_GPU_CONTROLLER_PATH=$CEDANA_GPU_CONTROLLER_PATH
 Environment=CEDANA_PROFILING_ENABLED=$CEDANA_PROFILING_ENABLED
 Environment=CEDANA_OTEL_ENABLED=$CEDANA_OTEL_ENABLED
 Environment=CEDANA_IS_K8S=$CEDANA_IS_K8S
