+++ cmd/flags.go
@@ -1,0 +1,19 @@
+package cmd
+
+// This file contains all the flags used in the cmd package
+
+const (
+	dirFlag            = "dir"
+	imgFlag            = "image"
+	idFlag             = "id"
+	rootFlag           = "root"
+	wdFlag             = "working-dir"
+	containerNameFlag  = "container-name"
+	containerRootFlag  = "container-root"
+	bundleFlag         = "bundle"
+	consoleSocketFlag  = "console-socket"
+	detachFlag         = "detach"
+	isK3sFlag          = "is-k3s"
+	netPidFlag         = "net-pid"
+	tcpEstablishedFlag = "tcp-established"
+)
